refactor!: rename module and delete old
